Fix checkptr-illegal uintptr→Pointer conversions and run -race in CI.
sessionIDFromEvent now takes unsafe.Pointer so the SCM handler converts
EVENTDATA at the call boundary instead of a delayed uintptr round-trip.
tokenOwnerSID reads TOKEN_OWNER.Owner as a SID from the buffer (GetTokenUser
shape) instead of reconstructing a pointer from a uintptr.

TestWindowsPipeACLDeniesNonAdmin skips when the process is not elevated.
Windows CI runs go vet and CGO_ENABLED=1 go test -race; Linux tests use -race.

Closes #69.

+++ internal/runtime/session_windows_test.go
@@ -12,7 +12,7 @@ import (
 
 func TestParseSessionChangeLogonLogoff(t *testing.T) {
 	n := windows.WTSSESSION_NOTIFICATION{Size: 8, SessionID: 3}
-	ptr := uintptr(unsafe.Pointer(&n))
+	ptr := unsafe.Pointer(&n)
 
 	sc, ok := ParseSessionChange(svc.SessionChange, windows.WTS_SESSION_LOGON, ptr)
 	if !ok || !sc.Logon || sc.SessionID != 3 {
